feat: split fetchers
