+++ packages/tarball-fetcher/src/index.ts
@@ -1,24 +1,20 @@
-import path from 'path'
 import PnpmError from '@pnpm/error'
 import {
   Cafs,
-  DeferredManifestPromise,
   FetchFunction,
   FetchOptions,
-  FetchResult,
 } from '@pnpm/fetcher-base'
 import {
   FetchFromRegistry,
   GetCredentials,
   RetryTimeoutOptions,
 } from '@pnpm/fetching-types'
-import gfs from '@pnpm/graceful-fs'
-import ssri from 'ssri'
 import createDownloader, {
   DownloadFunction,
   TarballIntegrityError,
-  waitForFilesIndex,
-} from './createDownloader'
+} from './remoteTarballFetcher'
+import createLocalTarballFetcher from './localTarballFetcher'
+import createGitHostedTarballFetcher, { waitForFilesIndex } from './gitHostedTarballFetcher'
 
 export { BadTarballError } from './errorTypes'
 
@@ -32,17 +28,22 @@ export default function (
     retry?: RetryTimeoutOptions
     offline?: boolean
   }
-): { tarball: FetchFunction } {
+): { localTarball: FetchFunction, remoteTarball: FetchFunction, gitHostedTarball: FetchFunction } {
   const download = createDownloader(fetchFromRegistry, {
     retry: opts.retry,
     timeout: opts.timeout,
   })
+
+  const remoteTarballFetcher = fetchFromTarball.bind(null, {
+    download,
+    getCredentialsByURI: getCredentials,
+    offline: opts.offline,
+  }) as FetchFunction
+
   return {
-    tarball: fetchFromTarball.bind(null, {
-      download,
-      getCredentialsByURI: getCredentials,
-      offline: opts.offline,
-    }) as FetchFunction,
+    localTarball: createLocalTarballFetcher(),
+    remoteTarball: remoteTarballFetcher,
+    gitHostedTarball: createGitHostedTarballFetcher(remoteTarballFetcher),
   }
 }
 
@@ -63,13 +64,6 @@ async function fetchFromTarball (
   },
   opts: FetchOptions
 ) {
-  if (resolution.tarball.startsWith('file:')) {
-    const tarball = resolvePath(opts.lockfileDir, resolution.tarball.slice(5))
-    return fetchFromLocalTarball(cafs, tarball, {
-      integrity: resolution.integrity,
-      manifest: opts.manifest,
-    })
-  }
   if (ctx.offline) {
     throw new PnpmError('NO_OFFLINE_TARBALL',
       `A package is missing from the store but cannot download it in offline mode. The missing package may be downloaded from ${resolution.tarball}.`)
@@ -84,42 +78,4 @@ async function fetchFromTarball (
     onStart: opts.onStart,
     registry: resolution.registry,
   })
-}
-
-const isAbsolutePath = /^[/]|^[A-Za-z]:/
-
-function resolvePath (where: string, spec: string) {
-  if (isAbsolutePath.test(spec)) return spec
-  return path.resolve(where, spec)
-}
-
-async function fetchFromLocalTarball (
-  cafs: Cafs,
-  tarball: string,
-  opts: {
-    integrity?: string
-    manifest?: DeferredManifestPromise
-  }
-): Promise<FetchResult> {
-  try {
-    const tarballStream = gfs.createReadStream(tarball)
-    const [fetchResult] = (
-      await Promise.all([
-        cafs.addFilesFromTarball(tarballStream, opts.manifest),
-        opts.integrity && (ssri.checkStream(tarballStream, opts.integrity) as any), // eslint-disable-line
-      ])
-    )
-    return { filesIndex: fetchResult }
-  } catch (err: any) { // eslint-disable-line
-    const error = new TarballIntegrityError({
-      attempts: 1,
-      algorithm: err['algorithm'],
-      expected: err['expected'],
-      found: err['found'],
-      sri: err['sri'],
-      url: tarball,
-    })
-    error['resource'] = tarball
-    throw error
-  }
 }
